Improved cache-miss-update efficiency.
Now, all incoming Samples are passed through a method which checks their
pre-requisites against the cache. Any missing pre-requisites are
fulfilled within that same method. That method, and hence all following
Samples, are blocked until the current Sample is complete.

This means, among other things, that for a block of Samples all devoted
to the same render-ID, we will only ever retrieve the WorldDescriptor
once, only inflate that into a World once, etc.